Make report publisher a deployer subcommand
