+++ deployer/__main__.py
@@ -9,6 +9,7 @@ import deployer.generate.generate_gcp_cluster  # noqa: F401
 import deployer.grafana.central_grafana  # noqa: F401
 import deployer.grafana.grafana_tokens  # noqa: F401
 import deployer.keys.decrypt_age  # noqa: F401
+import deployer.report # noqa: F401
 
 from .cli_app import app
